backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic-settings==2.5.0
pyjwt==2.9.0
pyyaml==6.0.2
anthropic==0.39.0
python-multipart==0.0.12
openpyxl==3.1.5
httpx==0.27.0
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0
